Move action attempt errors under WaitForActionAttempt module
