+++ spec/seam_client/wait_for_action_attepmt_spec.rb
@@ -119,7 +119,7 @@ RSpec.describe Seam::Http do
           action_attempt_id: action_attempt.action_attempt_id,
           wait_for_action_attempt: {timeout: 0.1}
         )
-      end.to raise_error(Seam::Errors::ActionAttemptTimeoutError) do |error|
+      end.to raise_error(Seam::Http::WaitForActionAttempt::TimeoutError) do |error|
         expect(error.action_attempt.action_attempt_id).to eq(action_attempt.action_attempt_id)
         expect(error.action_attempt.status).to eq(action_attempt.status)
       end
@@ -142,7 +142,7 @@ RSpec.describe Seam::Http do
           action_attempt_id: action_attempt.action_attempt_id,
           wait_for_action_attempt: true
         )
-      end.to raise_error(Seam::Errors::ActionAttemptFailedError) do |error|
+      end.to raise_error(Seam::Http::WaitForActionAttempt::FailedError) do |error|
         expect(error.message).to include("Failed")
         expect(error.action_attempt.action_attempt_id).to eq(action_attempt.action_attempt_id)
         expect(error.action_attempt.status).to eq("error")
@@ -167,7 +167,7 @@ RSpec.describe Seam::Http do
           action_attempt_id: action_attempt.action_attempt_id,
           wait_for_action_attempt: {timeout: 0.5, polling_interval: 3}
         )
-      end.to raise_error(Seam::Errors::ActionAttemptTimeoutError) do |error|
+      end.to raise_error(Seam::Http::WaitForActionAttempt::TimeoutError) do |error|
         expect(error.action_attempt.action_attempt_id).to eq(action_attempt.action_attempt_id)
         expect(error.action_attempt.status).to eq(action_attempt.status)
       end
